Move Gadget and Valid back
